Reuse code for Metal and WGSL entry point legalization (#6063)
* Refactor to reuse common for metal and wgsl entry point legalization

* refactor system val work item

* refactor simplify user names

* clean up fix semantic field of struct

* improve code layout

* split wgsl/metal to seperate classes and cleanup

* remove extra includes

* remove dead code comments

* minor cleanup

* squash merge from master and resolve conflict

* apply metal spec const thread count changes

* Revert "apply metal spec const thread count changes"

This reverts commit c42d707fd25ee0328598650d3235cd2322810ccc.

* Revert "squash merge from master and resolve conflict"

This reverts commit 06db88ef7001bdfe93fb23af35af0d026b255dee.

* Merge remote-tracking branch 'origin/master'

* apply metal spec const thread count changes

* Revert "apply metal spec const thread count changes"

This reverts commit 3b9e6f53cee2e6076ac2b7a0d015a1ed2cbbd627.

* Revert "Merge remote-tracking branch 'origin/master'"

This reverts commit 99869d573a46dadeb24445405f5a1e37a8e03d0d.

* apply metal spec const thread count changes

---------
